Clean up TinyIoC changes
Move AutoMockingRegistrationProvider into separate file, prefer TinyIoC
namespace, clarify enum names, add missing property to copy method in
ResolveOptions
