+++ SpecEasy/GenericSpec.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Rhino.Mocks;
 using Rhino.Mocks.Interfaces;
+using TinyIoC;
 
 namespace SpecEasy
 {
@@ -71,22 +72,6 @@ namespace SpecEasy
         {
             get { return Get<TUnit>(); }
             set { Set(value); }
-        }
-    }
-
-    internal class AutoMockingRegistrationProvider : IFallbackRegistrationProvider
-    {
-        public bool TryRegister(Type registerType, TinyIoCContainer container)
-        {
-            var mock = TryMockType(registerType);
-            if (mock == null) return false;
-            container.Register(registerType, mock);
-            return true;
-        }
-
-        private static object TryMockType(Type type)
-        {
-            return type.IsInterface || type.IsAbstract ? MockRepository.GenerateMock(type, new Type[0]) : null;
         }
     }
 }
